:truck: (conpositions/matrix.ts) move directory compositions to functions
